+++ services/assistant/app/adapters/output/grpc/__init__.py
@@ -1,0 +1,13 @@
+from app.adapters.output.grpc.client_factories import (
+    XrayClientFactory,
+    BookingClientFactory,
+)
+from app.adapters.output.grpc.xray_client import GRPCXrayClient
+from app.adapters.output.grpc.booking_client import GRPCBookingClient
+
+__all__ = [
+    "BookingClientFactory",
+    "XrayClientFactory",
+    "GRPCXrayClient",
+    "GRPCBookingClient",
+]
