+++ services/assistant/app/adapters/output/grpc/booking_client.py
@@ -1,9 +1,9 @@
 from app.contracts.clients import BookingClient
 
-from app.infra.grpc.gen.booking_pb2_grpc import BookingStub
-from app.infra.grpc.gen.booking_pb2 import BookingRequest, BookingResponse
+from app.adapters.output.grpc.gen.booking_pb2_grpc import BookingStub
+from app.adapters.output.grpc.gen.booking_pb2 import BookingRequest, BookingResponse
 
-from app.infra.grpc.grpc_client import GRPCClient
+from app.adapters.output.grpc.grpc_client import GRPCClient
 
 
 class GRPCBookingClient(GRPCClient, BookingClient):
