+++ services/assistant/app/adapters/output/grpc/grpc_client.py
