+++ services/assistant/app/adapters/output/grpc/xray_client.py
@@ -1,9 +1,9 @@
 from app.contracts.clients import XrayClient
 
-from app.infra.grpc.gen.xray_pb2_grpc import XrayStub
-from app.infra.grpc.gen.xray_pb2 import RestartResponse, Null
+from app.adapters.output.grpc.gen.xray_pb2_grpc import XrayStub
+from app.adapters.output.grpc.gen.xray_pb2 import RestartResponse, Null
 
-from app.infra.grpc.grpc_client import GRPCClient
+from app.adapters.output.grpc.grpc_client import GRPCClient
 
 
 class GRPCXrayClient(GRPCClient, XrayClient):
