--- services/assistant/app/infra/grpc/__init__.py
@@ -1,10 +1,0 @@
-from app.infra.grpc.client_factories import XrayClientFactory, BookingClientFactory
-from app.infra.grpc.xray_client import GRPCXrayClient
-from app.infra.grpc.booking_client import GRPCBookingClient
-
-__all__ = [
-    "BookingClientFactory",
-    "XrayClientFactory",
-    "GRPCXrayClient",
-    "GRPCBookingClient",
-]
